The Great Cleanup
Drop unused files. Drop large blocks of commented lines. Restore
from Git if needed in the future.

Drop projects/ as it is not referred by anything else.

Drop repos and build files if they are available in hdl/bazel_rules_hdl.

This prepares us for a future change to incorporate bazel_rules_hdl.