CI: always store and upload cache at the end
Github Actions skipped cache uploading when there is a perfect key
match. This works well for other types of projects (e.g. dependency
cache), but not us, as we cache the compiled objects of our sources.

The cache should updated whenever the sources change, so this change
includes the current commit hash in the cache key, while generic key
is used as "restore-key".