bazel: replace new_git_repository with http_archive
new_git_repository could be flaky due to various reasons, as it
uses the "git" utility on the host.

Plus, it is faster to download a compressed archive.

It is standard practice to use http_archive wherever possible. As
seen from other Bazel projects and GitHub code search results,
uses of git_repository are quite rare compared to http_archive.

Ref: https://github.com/tensorflow/tensorflow/commit/80acdaf4975b6572aa8dacd0e10997e17d370f69